autoloop.lua: properly detect and reset "load-file" property
Previously, the detection of `was_loop` was made too early (during the script loading stage). It overlooked the fact that the "loop-file" property can be changed by auto profiles. Delay this detection until the file is loaded. (Hopefully at that time auto profiles are already applied.)

The next thing is properly undoing the "set loop-file inf". This is now done when the file has ended. A dirty hack is required because "end-file" happens along with "start-file" of the next file. Blindly resetting this property after auto profiles that apply to the next file have changed it is not desired. This may break, if the auto profiles change this property after the if statement has succeeded and before the set_property_native line, but seems to be unlikely to happen (in my tests).

Closes #1